// src/assets/types/patientsType.ts
export interface patientsType {
  id: number;
  firstName: string;
  lastName: string;
  fullName: string;
  nationalCode: string;
  phoneNumber: string;
  dateOfBirth: string;
  age: number;
  gender: number;
  genderDisplay: string;
  address: string;
  registrationDate: string;
  appointmentsCount: number;
}